modules/feedback/requirements.txt
pika>=1.3.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
websockets